prevent duplicate error logging
Remove log.Error statements before returning an error.
This lets the consumer of the function decide if the error should be logged
and were it should be logged to.

Previously some code paths also logged the error before returning it
while others didn't, making it impossible to always log errors but prevent duplicates.